Switch to expand-contract classes
Removed the modal implementation. It could have worked, but it was the
most optimal solution. Expanding and contracting the menu is the better
way to handle this.

I also added a little script to handle the class. Probably will move that
off into its own little file and put it inside onload().